fix(app, platforms): moves PLATFORM_ID and PROVIDER_ID to platforms
