+++ platforms/src/POAP/App-Bindings.ts
@@ -5,6 +5,6 @@ export class POAPPlatform implements Platform {
   getOAuthUrl(state: string): Promise<string> {
     throw new Error("Method not implemented.");
   }
-  platformId = "Poap";
-  path = "Poap";
+  platformId = "POAP";
+  path = "POAP";
 }
